deletes the correct tag now. Moved dialog to its own component
