+++ src/components/context/ContextSidebar.tsx
@@ -7,17 +7,15 @@ import {
   AccordionTrigger,
 } from "~/components/ui/accordion";
 import { ScrollArea } from "~/components/ui/scroll-area";
-import { useAppContext } from "~/store";
 import { TagsIcon } from "lucide-react";
 
-import { Button } from "../ui/button";
 import AllNotes from "./AllNotes";
 import Login from "./Login";
 import TagItem from "./TagItem";
 import TrashedNotes from "./TrashedNotes";
+import DeleteTagDialog from "../tags/DeleteTagDialog";
 
 export default function ContextSidebar() {
-  const { currentNote } = useAppContext();
 
   async function fetchTags() {
     const apiResponse = await listTags({});
@@ -52,6 +50,7 @@ export default function ContextSidebar() {
                 </div>
               </AccordionTrigger>
               <AccordionContent>
+                <DeleteTagDialog />
                 {data?.map((tag) => <TagItem key={tag.id} tag={tag} />)}
               </AccordionContent>
             </AccordionItem>
