+++ src/store/index.ts
@@ -30,8 +30,11 @@ interface State {
   settings: Settings;
   setSettings: (settings: Settings) => void;
 
-  confirmTagDelete: boolean;
-  setConfirmTagDelete: (confirmTagDelete: boolean) => void;
+  deleteTagDialog: boolean;
+  setDeleteTagDialog: (deleteTagDialog: boolean) => void;
+
+  deleteTagDialogId: number | undefined;
+  setDeleteTagDialogId: (deleteTagDialogId: number | undefined) => void;
 }
 
 export const useAppContext = create<State>()(
@@ -66,8 +69,11 @@ export const useAppContext = create<State>()(
       settings: {},
       setSettings: (settings) => set({ settings }),
 
-      confirmTagDelete: false,
-      setConfirmTagDelete: (confirmTagDelete) => set({ confirmTagDelete }),
+      deleteTagDialog: false,
+      setDeleteTagDialog: (deleteTagDialog) => set({ deleteTagDialog }),
+
+      deleteTagDialogId: undefined,
+      setDeleteTagDialogId: (deleteTagDialogId) => set({ deleteTagDialogId }),
     }),
     {
       name: "captains-log-storage",
